Understand and tolerate `tailored=false` requests
Part of the changes aim to tolerate `tailored=false`, like in this url:

https://interactive.guim.co.uk/2016/05/brexit-companion/embed/embed.html?tailored=false&beginner=who_can_vote&intermediate=impact_economy&advanced=can_overrule_vote&default=beginner

...which can be produced by https://github.com/guardian/frontend/pull/13157,
which sets `tailored=true` to `tailored=false` for 50% of users.

The `id` value can still be used to set an atomId, but if tailored
params are present (ie `default`, `beginner`, `intermediate`, etc) they
will override the `id` value, even if `tailored=false` or is not even set.